python-scripts/t.py
# w3.eth.wait_for_transaction_receipt()
import binascii
import random
from utils.blockchain_utils import BlockchainUtils
from abis.weighted_pool_factory_abi import weighted_pool_factory_abi
from abis.vault_abi import vault_abi
from SETTINGS import *
import os
from dotenv import load_dotenv
import sys
import eth_abi
from hexbytes import HexBytes

load_dotenv()
PRK = os.getenv("ADMIN")

butils = BlockchainUtils(provider_url=RPC_URL, prk=PRK)
vault_contract = butils.get_contract(contract_address=VAULT, contract_abi=vault_abi)

tx_hsh = "0x82f41c98727021c228da98ecd5f45233f59c5df174426696d150a3e7dc630008"
tx_receipt = butils.w3.eth.wait_for_transaction_receipt(tx_hsh)

# look in logs for 
