python-scripts/requirements.txt
web3==6.16.0
requests
python-dotenv
